requirements.txt
audioread==2.1.9
hydra-core==1.1.1
joblib==1.0.1
librosa==0.8.1
matplotlib==3.4.3
nnmnkwii==0.1.1
numpy==1.22.4
omegaconf==2.1.1
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.0
pylint==2.13.9
pysptk==0.1.18
pyworld==0.3.0
PyYAML==5.4.1
scikit-learn==1.0
scipy==1.7.1
setuptools==58.3.0
SoundFile==0.10.3.post1
tensorboard==2.6.0
torch==1.12.1
torchaudio==0.12.1
tqdm==4.62.3